requirements.txt
pydantic==2.11.7
pydantic-settings==2.11.0
httpx==0.28.1
redis==6.2.0
pandas==2.3.1
numpy==2.3.2
alpaca-py==0.42.1
python-dotenv==1.1.1
websockets>=13.0
